requirements.txt
bcrypt==4.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
reflex==0.4.9
